test: split multi-day hike coverage
